requirements.txt
Django==1.7.1
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
South==1.0.1
wsgiref==0.1.2
